+++ test/fixtures/next-project/polyfills.js
@@ -1,24 +1,36 @@
+// @ts-check
+
 "use strict";
 
 if (typeof window === "undefined") {
-  if (!("performance" in global))
-    global.performance = require("perf_hooks").performance;
+  if (!("performance" in globalThis))
+    // @ts-ignore This isn’t a perfect polyfill.
+    globalThis.performance = require("perf_hooks").performance;
 
-  if (!("EventTarget" in global))
-    global.EventTarget =
-      require("events").EventTarget || require("event-target-shim").EventTarget;
+  if (!("EventTarget" in globalThis))
+    globalThis.EventTarget = require("event-target-shim").EventTarget;
 
-  if (!("Event" in global))
-    global.Event =
-      require("events").Event || require("event-target-shim").Event;
+  if (!("Event" in globalThis))
+    globalThis.Event = require("event-target-shim").Event;
 
-  if (!("CustomEvent" in global))
-    global.CustomEvent = class CustomEvent extends Event {
-      constructor(eventName, { detail, ...eventOptions } = {}) {
-        super(eventName, eventOptions);
-        this.detail = detail;
-      }
-    };
+  if (!("CustomEvent" in globalThis))
+    // @ts-ignore This isn’t a perfect polyfill.
+    globalThis.CustomEvent =
+      /**
+       * @template [T=unknown]
+       * @type {globalThis.CustomEvent<T>}
+       */
+      class CustomEvent extends Event {
+        /**
+         * @param {string} eventName
+         * @param {object} [options]
+         * @param {T} [options.detail]
+         */
+        constructor(eventName, { detail, ...eventOptions } = {}) {
+          super(eventName, eventOptions);
+          if (detail) this.detail = detail;
+        }
+      };
 
   require("abort-controller/polyfill");
 }