+++ test/fsPathRemove.mjs
@@ -1,12 +1,11 @@
+// @ts-check
+
 import execFilePromise from "./execFilePromise.mjs";
 
 /**
  * Removes a file or directory.
- * @kind function
- * @name fsPathRemove
  * @param {string} path Filesystem path.
- * @returns {Promise<void>} Resolves once the file or directory is removed.
- * @ignore
+ * @returns Resolves once the file or directory is removed.
  */
 export default async function fsPathRemove(path) {
   if (typeof path !== "string")