+++ test/startNext.mjs
@@ -1,3 +1,5 @@
+// @ts-check
+
 import { createServer } from "http";
 import { createRequire } from "module";
 
@@ -6,12 +8,10 @@ import listen from "./listen.mjs";
 const require = createRequire(import.meta.url);
 
 /**
- * Starts [Next.js](https://nextjs.org).
- * @kind function
- * @name startNext
- * @param {string} [dir] [Next.js](https://nextjs.org) project directory path.
- * @returns {Promise<{port: number, close: Function}>} Resolves the port the server is listening on, and a function to close the server.
- * @ignore
+ * Starts Next.js.
+ * @param {string} dir Next.js project directory path.
+ * @returns Resolves the port the server is listening on, and a function to
+ *   close the server.
  */
 export default async function startNext(dir) {
   const next = require(require.resolve("next", { paths: [dir] }));